Version 21/10/24

+++ src/main/java/com/BotCervecerias/Repositories/UserRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface UserRepository  extends CrudRepository<Users,Integer>{
+public interface UserRepository  extends CrudRepository<Users,Long>{
     Optional<Users>findByEmail(String email);
 }
